requirements.txt
langchain
llama-index
haystack
fastapi
uvicorn
requests
python-dotenv
pandas==2.0.3
langchain-groq